# backend/openapi.yml
openapi: 3.1.0
info:
  version: '1'
  title: 'MyAce API Documentation'
  x-logo:
    url: 'https://myace.ai/logo.svg'
    # backgroundColor: '#FFFFFF'
    altText: 'MyAce logo'
servers:
  - url: 'https://api.myace.ai'
paths:
...
  /enterprises:
    post:
      summary: Create an enterprise
      description: This route creates a new enterprise. The authenticated user making this request must be a part of the MyAce team.
      tags:
        - Enterprises
      requestBody:
        required: true
        content:
          'application/json':
              schema:
                $ref: "#/components/schemas/NewEnterprise"
      responses:
        '201':
          description: Enterprise created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        default:
          $ref: '#/components/responses/DefaultError'
  /enterprises/{enterprise_id}:
    get:
      summary: Retrieve an enterprise
      description: This route retrieves an enterprise by its unique ID.
      tags:
        - Enterprises
      parameters:
        - name: enterprise_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EnterpriseID'
      responses:
        '200':
          description: Retrieved enterprise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enterprise"
        default:
          $ref: '#/components/responses/DefaultError'

    patch:
      summary: Update an enterprise
      description: This route edits an enterprise by its unique ID.
      tags:
        - Enterprises
      parameters:
        - name: enterprise_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EnterpriseID'
      requestBody:
        description: All fields are optional
        required: false
        content:
          'application/json':
              schema:
                $ref: "#/components/schemas/UpdateEnterprise"
      responses:
        '200':
          description: Updated enterprise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enterprise"
        default:
          $ref: '#/components/responses/DefaultError'

    delete:
      summary: Delete an enterprise
      description: This route permanently deletes the specified enterprise.
      tags:
        - Enterprises
      parameters:
        - name: enterprise_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/EnterpriseID'
      responses:
        '204':
          description: Enterprise deleted
        default:
          $ref: '#/components/responses/DefaultError'
  /enterprises/{enterprise_id}/invitations:
    post:
      summary: Invite user to enterprise
      description: This route invites a user by email to an enterprise. If the user account does not already exist, registration instructions will be sent to their email.
      tags:
        - Enterprise Invitations
      requestBody:
        required: true
        content:
          'application/json':
              schema:
                type: object
                required: [user_email, role]
                properties:
                  user_email:
                    $ref: "#/components/schemas/UserEmail"
                  role:
                    $ref: "#/components/schemas/EnterpriseRole"
      responses:
        '201':
          description: Enterprise invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationForEnterprise'
        default:
          $ref: '#/components/responses/DefaultError'

    get:
      summary: Retrieve all outgoing invitations
      description: This route retrieves all outgoing enterprise invitations for an enterprise manager to view.
      tags:
        - Enterprise Invitations
      responses:
        '200':
          description: Retrieved outgoing invitations
          content:
            application/json:
              schema:
                type: object
                required: [invitations]
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: "#/components/schemas/InvitationForEnterprise"
        default:
          $ref: '#/components/responses/DefaultError'
  /enterprises/invitations:
    get:
      summary: Retrieve all incoming invitations
      description: This route retrieves all incoming enterprise invitations for the recipient to view.
      tags:
        - Enterprise Invitations
      responses:
        '200':
          description: Retrieved incoming invitations
          content:
            application/json:
              schema:
                type: object
                required: [invitations]
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: "#/components/schemas/InvitationForRecipient"
        default:
          $ref: '#/components/responses/DefaultError'
  /enterprises/invitations/{invite_id}:
    delete:
      summary: Delete invitation
      description: This route deletes an invitation to join an enterprise. Calling this route means that an enterprise manager has uninvited a user, or a user has rejected an incoming invitation.
      tags:
        - Enterprise Invitations
      parameters:
        - name: invite_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EnterpriseInviteID"
      responses:
        '204':
          description: Deleted/Rejected invite
        default:
          $ref: '#/components/responses/DefaultError'
  /enterprises/invitations/{invite_id}/accept:
    post:
      summary: Accept incoming invitation
      description: This route accepts an incoming invitation to join an enterprise.
      tags:
        - Enterprise Invitations
      parameters:
        - name: invite_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/EnterpriseInviteID"
      responses:
        '204':
          description: Accepted invite
        default:
          $ref: '#/components/responses/DefaultError'

  # /users/forgot/:
  #   post:
  #     summary: Forgot password
  #     description: This route sends a password recovery link to a specified email if the email belongs to a user. If the provided email is invalid, no email will be sent, but the response will still indicate success in order to mitigate abuse.
  #     tags:
  #       - TODO # TODO
  #     requestBody:
  #       required: true
  #       content:
  #         'application/json':
  #             schema:
  #               type: object
  #               required:
  #                 - email
  #               properties:
  #                 email:
  #                   $ref: "#/components/schemas/UserEmail"
  #     security: [] # no authentication
  #     responses:
  #       '204':
  #         description: Email possibly sent.
  #       default:
  #         $ref: '#/components/responses/DefaultError'
components:
  responses:
    DefaultError:
      description: Unexpected error
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
  schemas:
...
    NewEnterprise:
      type: object
      required: [name]
      properties:
        name:
          $ref: "#/components/schemas/EnterpriseName"
        website:
          $ref: "#/components/schemas/EnterpriseWebsite"
        support_email:
          $ref: "#/components/schemas/EnterpriseEmail"
        support_phone:
          $ref: "#/components/schemas/EnterprisePhoneNumber"
    UpdateEnterprise:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/EnterpriseName"
        website:
          $ref: "#/components/schemas/EnterpriseWebsite"
        support_email:
          $ref: "#/components/schemas/EnterpriseEmail"
        support_phone:
          $ref: "#/components/schemas/EnterprisePhoneNumber"
    Enterprise:
      type: object
      required:
        - enterprise_id
        - name
        - website
        - support_email
        - support_phone
        - logo
        - created_at
      properties:
        enterprise_id:
          $ref: "#/components/schemas/EnterpriseID"
        name:
          $ref: "#/components/schemas/EnterpriseName"
        website:
          allOf:
            - nullable: true
            - $ref: "#/components/schemas/EnterpriseWebsite"
        support_email:
          allOf:
            - nullable: true
            - $ref: "#/components/schemas/EnterpriseEmail"
        support_phone:
          allOf:
            - nullable: true
            - $ref: "#/components/schemas/EnterprisePhoneNumber"
        logo:
          allOf:
            - nullable: true
            - $ref: "#/components/schemas/EnterpriseLogo"
        created_at:
          type: string
          format: date-time
    InvitationForEnterprise:
      type: object
      description: An invitation to an enterprise from the perspective of that enterprise's manager.
      required:
        - invite_id
        - user_email
        - role
        - created_at
      properties:
        invite_id:
          $ref: "#/components/schemas/EnterpriseInviteID"
        user_email:
          $ref: "#/components/schemas/UserEmail"
        role:
          $ref: "#/components/schemas/EnterpriseRole"
        created_at:
          type: string
          format: date-time
    InvitationForRecipient:
      type: object
      description: An invitation to an enterprise from the perspective of the invitee.
      required:
        - invite_id
        - role
        - enterprise
        - created_at
      properties:
        invite_id:
          $ref: "#/components/schemas/EnterpriseInviteID"
        role:
          $ref: "#/components/schemas/EnterpriseRole"
        enterprise:
          $ref: "#/components/schemas/Enterprise"
        created_at:
          type: string
          format: date-time


    # ------------------------------ Common types ------------------------------
...
    UserEmail:
      type: string
      description: The user's email address.
      format: email
      pattern: ^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$
      example: johnsmith@gmail.com
...
    EnterpriseID:
      type: string
      description: The enterprise's unique identifier.
      format: uuid
    EnterpriseName:
      type: string
      description: The public name of an enterprise.
      example: Laguna Beach Tennis Club
    EnterpriseWebsite:
      type: string
      description: The homepage of an enterprise.
      example: https://www.lagunabeachtennisclub.com
    EnterprisePhoneNumber:
      type: string
      description: The support phone number for an enterprise.
      pattern: ^\d+$
    EnterpriseEmail:
      type: string
      description: The support email address for an enterprise.
      format: email
      pattern: ^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$
      example: support@lagunabeachtennisclub.com
    EnterpriseLogo:
      type: string
      description: The URL that points to an enterprise's 1:1 logo image.
      example: https://logo.png
    EnterpriseRole:
      type: string
      description: The role of a user within an enterprise.
      enum:
        - parent
        - player
        - instructor
        - admin
    EnterpriseInviteID:
      type: string
      description: The enterprise invitation's unique identifier.
      format: uuid
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
security:
  - bearerAuth: []